Create new SelectLayer button at the right top of the map
