+++ src/components/Map/MapComponent.css
@@ -9,7 +9,7 @@
   box-sizing: border-box;
   border-width: 2px;
   border-color: rgb(255 255 255);
-  background-color: rgb(59 130 246);
+  background-color: rgb(255 255 255);
 }
 
 .Map-popup-title {
