+++ src/components/SelectLayer/SelectLayer.css
@@ -1,6 +1,5 @@
 .SelectLayer {
   user-select: none;
-  display: flex;
 }
 .SelectLayer-inner {
   /*min-height: 100px;*/
